進捗:ニュートラルを追加

+++ second_swerve/v1.0/src/func.h
@@ -1,3 +1,3 @@
 #pragma once 
 
-int16_t fmap(double x, double in_min, double in_max, int16_t out_min, int16_t out_max);
+int fmap(double x, double in_min, double in_max, int out_min, int out_max);
